better error reporting in run

+++ runforrest.py
@@ -191,6 +191,11 @@ class Executor:
                 time.sleep(0.1)
 
     def _finish_task(self, file):
+        if not (self.done_dir / file).exists():
+            (self.todo_dir / file).rename(self.fail_dir / file)
+            print(f'task {file} failed without error.')
+            return
+
         (self.todo_dir / file).unlink()
         with open(self.done_dir / file, 'rb') as f:
             task = dill.load(f)
@@ -198,6 +203,7 @@ class Executor:
         if process.returncode == 0:
             yield task._value
         else:
+            print(f'task {file} failed with error {task._error}.')
             (self.done_dir / file).rename(self.fail_dir / file)
 
     def todo_tasks(self):
